@@ -3,8 +3,9 @@ package wseemann.media.romote.activity;
 import android.appwidget.AppWidgetManager;
 import android.content.Intent;
 import android.os.Bundle;
-import android.support.v4.app.FragmentTransaction;
-import android.support.v7.app.AppCompatActivity;
+
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.fragment.app.FragmentTransaction;
 
 import wseemann.media.romote.R;
 import wseemann.media.romote.fragment.MainFragment;
