@@ -3,7 +3,8 @@ package wseemann.media.romote.activity;
 import android.content.SharedPreferences;
 import android.os.Bundle;
 import android.preference.PreferenceManager;
-import android.support.v7.app.AppCompatActivity;
+
+import androidx.appcompat.app.AppCompatActivity;
 
 import com.jaku.core.JakuRequest;
 import com.jaku.core.KeypressKeyValues;
