@@ -1,14 +1,13 @@
 package wseemann.media.romote.utils;
 
 import android.app.Notification;
-import android.app.NotificationChannel;
-import android.app.NotificationManager;
 import android.app.PendingIntent;
 import android.content.Context;
 import android.content.Intent;
 import android.graphics.Bitmap;
 import android.os.Build;
-import android.support.v4.app.TaskStackBuilder;
+
+import androidx.core.app.TaskStackBuilder;
 
 import com.jaku.core.KeypressKeyValues;
 
